test: refactor egress nodes into seperate file
